+++ src/lang_utils/error_codes.ml
@@ -230,7 +230,8 @@ let warning_codes = [
   "M0217", None, "Redundant `persistent`";
   "M0218", None, "Redundant `stable`";
   "M0222", None, "Ignored `async*`";
-]
+  "M0223", None, "Redundant type instantiation";
+  ]
 
 let try_find_explanation code =
   match List.find_opt (fun (c, _) -> String.equal c code) error_codes with